Update patch set 1
Patch Set 1: Code-Review-2

Patch 1 is an initial proof of concept design. I think it captures in general what we want it to do however one reason I kind of don't like it is that it's a bash script so while will work on Linux and Mac. Windows developers won't be able to use it.

I was thinking of perhaps converting it to Python since I saw another ODL project use it for their scripts to launch mininet instances I thought it'd be reasonable to assume a developer might have Python installed.

Thoughts?

Patch-set: 1
Label: Code-Review=-2