Update patch set 1
Abandoned

After some investigation I've realized "jenkins-jobs test" mode already does a good enough job as a basic style checker. For certain fields such as job-templates and defaults it actually verifies that these exist and will fail if you provide an invalid name. As far as I can tell fields that it does not restrict / verify are fields that can be variable in value and wouldn't be able to figure out what the variable is so understandable.

I think we don't need to do much more on top of the test mode until proven otherwise.

Patch-set: 1
Status: abandoned