Update patch set 4
Patch Set 4:

> My one question is can we find a better way to store common Maven goals and options in a central place.

Unfortunately I can't make it a macro since JJB doesn't allow "project types" to be macros and the Maven related configurations are general project type config.

One option is to make 2 sets of defaults. One defined in releng-defaults.yaml for example. {mvn-goals-default} and {mvn-opts-default} which has the default settings and use those variables in place of our current MAVEN_GOALS and MAVEN_OPTS.

Then inside the project.yaml file we also define {project-mvn-goals} and {project-mvn-opts} which the project will need to switch if they don't want to use the releng-defaults one. They of course will also have to remove the first line comment as this would be considered a modification to the template. If we go this route though it I think it would make the enhancement bug more difficult to do (https://bugs.opendaylight.org/show_bug.cgi?id=2411) because now we have to consider if they are using the mvn settings from releng-defaults or the one from their project.yaml file.

Note: JJB does not allow multiple properties of the same type to have the same name.

Patch-set: 4
Label: Verified=0