Update patch set 2
Patch Set 2:

Dana,

I added you to this code review because this could affect BGPCEP if you want to use this feature. I used BGPCEP as an example of what the configuration format could look like if we want to allow projects to customize some variables while still using the base templates for the jobs.

Colin, Dana,

Please review this proposal and if you think it's a good idea please +1 it to let me know you're ok with this change to releng/builder.

Patch-set: 2
Label: Verified=0