Add the "basic builder" image definition
This adds the basic-builder Vagrant image definition. It is currently
crafted to specifically be an EL7 image as some of the libraries needed
for building OpenDOVE are not available in publicly accessible yum repos
as the packages built by The Linux Foundation to support building
OpenDOVE conflict with upstream packages. As such, they are not being
released unless specifically requested for EL6 systems.

Change-Id: Ia6464091780c820e4900eb224de845dcb072a272
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>