Add integration job to common template
This patch adds a new job type project-integration to the common job
template. It serves to be the single job to handle all integration jobs
existing on the master silo. This integration job is by default disabled
and unused.

Projects can pass a list of dependent projects in their project.cfg file for example:

DEPENDENCIES: aaa,controller,yangtools

This will flip the switch to enable the project-integration job and
cause a build to be triggered when any of the dependent jobs are built.

For email it will send an email with the project tags for ALL projects
listed in the dependencies list thusly notifying all dependent projects
of the build issue.

Change-Id: I8e06fb378dadb6d8b083b08cba366eb863efac76
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>