Update patch set 2
Patch Set 2:

FWIW I just built groupbasedpolicy on my laptop with Maven 3.2.5 and it was successful so I suspect whatever caused issues before is now resolved?

As for the FindBugs thing. I think modifying this would cause your project to no longer be able to take advantage of the templates we provide.

I propose an alternate solution if enabling FindBugs checkbox does not cause issues with projects that don't have it to build. I propose we add it to the templates instead just like we did with JaCoCo in a separate patch. This will let projects that use it take advantage of it and projects that don't just have it ignored (assuming it doesn't fail the build if FindBugs isn't available). How does that sound?

Patch-set: 2