DO NOT MERGE Very basic pcep job using pcc-mock.
This is a draft, it does not fully conform to Integration
job guidlines, and functionality is very basic.

What it does: Download pcc-mock, connect to
and disconnect from controller, watching for pcep-topology.
Link to latest run on Sandbox:
https://jenkins.opendaylight.org/sandbox/job/bgpcep-pccmock-master/30/

Future plan is to include json normalization library,
move to appropriate directories and get rid of Sandbox specifics.
Later, expand tested scenarios.

Change-Id: I3ab7a4624bd2d7bdbb1b6b7677d7574cb81c31ab
Signed-off-by: Vratko Polak <vrpolak@cisco.com>