Update patch set 3
Patch Set 3:

I've just spun a new test image and hooked it up to the sandbox for the dynamic_mininet slave. I've triggered a CSIT job that's defined there that will spin it up and try it. If that passes I'll let you setup other jobs to test that cbench is operating as expected. If everything passes we can merge

Patch-set: 3
Label: Verified=0