Update patch set 3
Patch Set 3: Code-Review+1

OK, this doesn't break already created tests. JamO if you have a test you could poke into sandbox to verify cbench is working as expected we can move on to getting this merged and the slave connected to releng.

Provisional +1. It'll get +2 and verified +1 when I see a working test on sandbox that utilizes cbench.

Patch-set: 3
Label: Code-Review=+1