Configure Vagrants for both ORD & DFW
We're spinning up the ORD (Chicago) Rackspace cloud and we need to have
an easy way for causing the Vagrant to spin-up on the correct network.
Historically I've done this with modifications to my copy of the
Rackspace dummy box Vagrant file but this modification is more flexible.

Secondary modification is to switch from an rs.flavor of performance1-1
to general1-1 since Rackspace seems to be trying to switch to that
(performance isn't listed in the webUI but the equivalent is general)

Change-Id: I39adf5613c366f9008d0eb8e7455b6404f82518f
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>