Add the lf-networking vagrant overlay
While the information used in this vagrant is extremely reliant on the
networking configuration that The Linux Foundation uses for the build
environments for ODL nothing in it is overtly private. As such, it is
being put in with the rest of the vagrant management for the project so
that LF admins may easily share it and others may have a slightly better
idea of how the environment actually goes together.

Change-Id: Idbb82dfd15f6cfcb6c7ad2737a9896cb5acb07a7
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>