Configure integration scripts to use Nexus proxy
We have an environment variable of ODLNEXUSPROXY that sets the base URL
path for (everything before /content) for scripts that aren't using the
maven settings. This patch updates the integration raw includes to use
the enviroment variable if it's set and to default to
https://nexus.opendaylight.org as they've been previously hard coded
if it the environment variable is not set.

Change-Id: I6ba7b450674166a50cde58313744da9724179a08
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>