Update patch set 2
Patch Set 2:

Thanks for the comments, Daniel.  I'm hoping Andy or Thanh can veriy
if the python setup step will work, or has the "C compiler" error you are 
seeing.  I tested it on my EC2 instance (something fedora) and it was ok.

I can't speak to the iptables command.  Looks like that was there from the 
initial commit of this file.

Patch-set: 2
Label: Verified=0