Update patch set 4
Patch Set 4:

Looks like the *monkey patch* that was done to try this out in the sandbox has things working as expected.  I was able to get a job to run and the robot suite to finish.  It fails, but it's failing for the right reasons.  The perf number it saw was
not high enough.  Exactly what we'd expect to see.

https://jenkins.opendaylight.org/sandbox/job/openflowplugin-csit-1node-imds-cbench-performance-only-master/18/robot/report/report.html

I can do the work to move this to the vagrant files if/when needed.   For now I think this is where we want it.

Patch-set: 4
Label: Verified=0