Update patch set 1
Patch Set 1: Code-Review+1

This seems ok to me, BUT we may want to enable the Lithium redesign openflow feature.  Maybe we can get this in CI for now, and when we decide if we are going to use the openflowplugin-app-new-lldp-speaker feature we can change it again.

Patch-set: 1
Reviewer: Gerrit User 2264 <2264@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+1