Add Jenkins Job to run tempest against odl-neutron-dummyprovider-test
This gerrit introduces a daily job that will run stub neutron against
latest openstack's tempest. When Kilo becomes available, this should
be converted to use that instead.

In order to provide an arbritary karaf feature to be loaded with ODL,
the following gerrit in networking-odl is going to be needed:
   https://review.openstack.org/#/c/175213/

Change-Id: I95bd1f4c51e1f38d9f1337724ee6523ba4e468a6
Signed-off-by: Flavio Fernandes <ffernand@redhat.com>