Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

As I mention in my comment, this isn't the way to do this. If we go this route we'll need to get a list of all the ports we may need and actually add them to the firewall instead. That's honestly more trouble than just allowing tests to push / remove iptables rules.

Patch-set: 1
Label: Code-Review=-1