Keeping 6-ish months worth of plot data
the data that the plot plugin reads from is actually very small, and even with 5400 values I don't think it would be much of a blip.

If the plot plugin can scall to 5400 values then I think we can get close to recording 6 months worth of data points for the jobs that run on merges.

for the jobs that are marked periodic, I used 180 as the default since I assumed most periodic jobs might be daily and 180 days is about 6 months.

all of these values are just defaults and can be overridden by each job as they see fit.

Change-Id: I6b622fe1c45cb48cac25c81b77d7b317c1a2fd81
Signed-off-by: Jamo Luhrsen <james.luhrsen@hp.com>