Install python-pip on devstack slaves
The devstack slave nodes are failing again. The current tests look to be
failing because they don't have ansible installed, but that is not the
problem. They don't have python-pip installed which is being used to
install virtualenv (not relying on the system package) and then install
ansible into that environment.

Change-Id: I9d429925d8e5c24fae040107f2a23a45d111d2d8
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>