Update patch set 2
Patch Set 2:

It is up to you, just +1 when you're ready for us to merge this. The only downside to having the jobs there before stable/lithium is branched is that the daily type jobs will fail everyday until that branch is created and you will receive emails from jenkins about these failures.

Patch-set: 2
Label: Verified=0