Set a root password on the mininet systems
My investigations into netopeer make me think that a non-root user isn't
going to be able to use it properly. Let's force the root user account
to have a password of root for the mininet test system so that jenkins
can call into netconf as root instead of as the jenkins user.

Change-Id: Ie574a81f50f8d8f946e24a20981c02168b44537d
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>