Update patch set 1
Patch Set 1:

That may be, but our policy is that if the test lives in a particular project, then we get a +1 from one (or more) committers of that project. If it was in jjb/integration then it would be an integration committer we would be needing. Because of the cross-project nature of this particular project we need to maintain the final ownership of tests that would have historically run in their private jenkins silo.

Patch-set: 1