Bug 3122: Allow JDKs to be enabled per branch in auto-jjb scripts
This patch proposes a change to the format of the configuration section
for branches as follows:

BRANCHES:
    - master:
        jdks: openjdk8
    - stable/helium:
        jdks: openjdk7
    - stable/lithium:
        jdks: openjdk7, openjdk8

The jdks setting will be a "required" field and the first JDK listed
will be the jdk that is used for all non-verify jobs.

- Removes support for the --jdks option as this now needs to be
  configured via CFG file.

As this formatting affects all projects and would require a
modification to all projects that use the BRANCHES configuration.

Change-Id: I09f52812b7cedd699ed7da9ffa3858ca4ed9037b
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>