Update patch set 1
Patch Set 1:

Hi YuLing.

The controller VMs should have Hbase installed as https://git.opendaylight.org/gerrit/#/c/19440 was merged.
This change will create a job which starts the suite from https://git.opendaylight.org/gerrit/#/c/18441 which attempts to start Hbase and do tests with it.
If the VMs were somehow not recreated after the merge of 19440 (or if the unpack was not an installation enough), the suite would show a lot of failures, but I would take it as a useful test result.
Overall, I am for "merge now, look at failures later" approach.

The alternative would be to ask Vasanthan for Sandbox job that proves the suite works.

Vratko.

Patch-set: 1