Update patch set 4
Patch Set 4:

Hi Vratko I agree with you however Venkat has just done what he saw we are doing in integration today. Moreover the work to split the shell scripts in pieces and recommend how to use the different pieces is full integration responsibility, so unless we come up with a good strategy to split the pieces, implement it in the jobs that we have today and have a good documentation for it in less than 24 hours, I will recommend to review and merge this patch as-is.

Patch-set: 4
Label: Verified=0