Update patch set 2
Patch Set 2: Code-Review+1

(1 comment)

Here is a link to job specification file, the resulting job was used for testing on Sandbox.
http://pastebin.com/iqwGJKyL

By default, that created job uses the start-vtn-coordinator action (to show it is run, but fails). When I removed it from "build with parameters" pre-filled form, the job went successfully (low passrate, but that is issue of bgpcep userfeatures tests in general, not affected by this change). Sandbox links for the two job runs:
https://jenkins.opendaylight.org/sandbox/job/bgpcep-csit-1node-cds-userfeatures-only-master/21/
https://jenkins.opendaylight.org/sandbox/job/bgpcep-csit-1node-cds-userfeatures-only-master/22/

From my point of view, patch set 2 is good enough to be committed, but perhaps you will want to test it some more to see it does not break anything.

Patch-set: 2