Update patch set 2
Patch Set 2:

I'm no expert to these things, but the gist of it all seems fruitful.  Especially if it's just adding more flexibility without changing how things work by default (that's how I understand it currently)

Things are growing in complexity with this stuff to the point that I feel it's very intimidating to newcomers.

And, since I have the stage, the other thing I want to point out for those jobs that end up running our robot suites, is that any of this ACTIONSAFTER stuff can easily be done there, and by default that's where I'll be doing stuff like that because that means I only have to muck with one project and the +2/merge that I have to get.  instead of putting robot code in the integration repo, then doing bash stuff in the builder repo.

anyway, nothing against what this patch is doing.  I can see it's benefits for sure.

Patch-set: 2
Reviewer: Gerrit User 2264 <2264@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0