Update patch set 2
Patch Set 2: Code-Review-1

So I tried this patch in the sandbox and it does not work because we do not have access to the releng/builder repo when executing a csit job. The way I see this could work is to store the shell scripts in the integration repository that we always download to do a test.

Patch-set: 2
Label: Code-Review=-1