Update patch set 2
Patch Set 2:

I don't think there is a way to use the same test plan file for different jobs.  seems the testplan must be named ${project}-${functionality} and if we try to keep those 
the same then the job name ends up being a duplicate and JJB rejects it.

Patch-set: 2