Update patch set 1
Patch Set 1:

Guys,

I don't know if this works.  In the sandbox this change is not taking affect.

However, I am able to change these two parameters (build-days-to-keep and build-num-to-keep) and see it reflected in the sandbox.  Doing the same for build-artifact-num-to-keep is still staying with the default (from releng-defaults.yaml) of 5.

will it work in production, or is there something I am missing?

Patch-set: 1
Label: Verified=0