Update patch set 10
Patch Set 10: Code-Review-1

I just realized that if you hardcode the VTNC version to be 6.2.0.0-SNAPSHOT, you will not be able to test VTNC releases. One idea to fix this is to run a command in robot to search foe the right version in the integration repository that is downloaded as part of the csit job. I can see the VTNC version is always stored in the integration file distributions/karaf/pom.xml

Patch-set: 10
Label: Code-Review=-1