Update patch set 2
Patch Set 2:

So the problem is I do not see a way to pass the rackspace vm name "rk-f21-mininet" as a variable from the jjb project down to the template. Thanh or Andy, any suggestion? if that is the case then every time we introduce a new OS (in mininet or controller) we need a new template. Only saving is to keep the template for OS upgrades, that is why I created patchset 2 (patchset 1 was f21 instead of fedora).

Patch-set: 2