Update patch set 1
Patch Set 1:

> Anyway people mostly fix bugs in stable/lithium now, so master can wait a bit in my opinion.

Stated this way, it does not sound like a good reason to break coherence between Lithium and post-Lithium testing.

On the other hand, the amount of jobs definitions in JJB folder is hopefully not high enough to cause this test job to be lost forever without anyone noticing (compared to amount of suites in Integration), and skipping the post-Lithium job execution until Lithium release is good for Releng Jenkins load.

Overall, I agree with not running on post-Lithium for now.

Patch-set: 1