Update the SELinux permissions for OVS
The SELinux port addition for OVS was missing an attribute and we
weren't seeing the reported error because of when and how the jenkins
spin-up scripts run. This fixes that

In addition, we take openvswitch into full SELinux permissive mode to
get around the fact that current SELinux policy does not allow it to
network connect to another system, even when connecting on an
openvswitch_port_t defined port.

Change-Id: I6c8d6240e090f3f723d454c61c511ed4af325c8d
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>