Update patch set 1
Patch Set 1: Code-Review-1

To be honest, I doubt this will actually correct the problem that you're seeing and may actually start causing issues for verify jobs. The reason is our problem is related to the JClouds plugin forgetting about slaves that it has started and therefore not stopping slaves it has never properly attached which means it's finding that the slave pools aren't emptying always. Integration goes through more slaves than any other project and therefore is _more_ likely to run into this issue, moving into the verify pool has a high chance of causing verifies to start having issues as well over time.

However, if you really feel that this is needed we can merge it, but I would really prefer not to.

Patch-set: 1