Update patch set 1
Patch Set 1:

This will cause these changes for VM controller runs in: 1. iptables service will be running; 2. sshpass will be not installed; 3. Hbase will be not installed; 4. default user will not have permission to sudo.

I suggest to revert this change as soon as a better fix is found.

Patch-set: 1