Update patch set 3
Patch Set 3: Code-Review-2

Thanks Vratko, just setting a -2 so that we don't accidently merge this. Next steps is to get agreement from the community that this is the direction we want to go as this change will affect all projects.

We should also come up with a script to rename the existing "master" jobs to Beryllium. I might be able to make a bash script using the Rest API to do this but I'll need some time to look into this.

Patch-set: 3
Label: Code-Review=-2