Update patch set 3
Patch Set 3:

> Once a project branches I expect them to create a project.cfg file and add the STREAM configuration.

Would not it be easier to require projects to create {project}.cfg file even before first jobs are enabled? Then we can get rid of the object, the whole -s argument, set sonar to the first STREAM branch and let the script fail if something was undefined?

Patch-set: 3