Update patch set 3
Patch Set 3:

> > Once a project branches I expect them to create a project.cfg
 > file and add the STREAM configuration.
 > 
 > Would not it be easier to require projects to create {project}.cfg
 > file even before first jobs are enabled? Then we can get rid of the
 > object, the whole -s argument, set sonar to the first STREAM branch
 > and let the script fail if something was undefined?

Originally the goal of the script was to make it as easy as possible for a new project so they'd just run the script with no parameters if they were a default project and didn't need any tweaks. I actually don't expect projects to use the script directly once they've created their CFG file. Considering the options are not really required any more by projects maybe we should use the script to initialize the default CFG file and recommend projects just edit that file once it's there.

Patch-set: 3