Update patch set 3
Patch Set 3:

> Anyway, I have noticed that job rename only goes through if the job
 > being renamed is not running (I am not sure what about if it is
 > pending, waiting for slave). So such a global rename operation can
 > be tricky to achieve on our active RelEng Jenkins.

So I tested on the sandbox and can confirm that a pending job can be renamed. Only running jobs cannot, which makes sense to me because a job that is in progress has a $WORKSPACE allocation which cannot be renamed while it is being used.

I guess the strategy here is to set Jenkins into Prepare to Shutdown mode and wait for all jobs to finish and new jobs to queue. Once no jobs are in flight run the rename script to rename all existing jobs before disabling the Shutdown mode.

Patch-set: 3