Add ofextensions/circuitsw project macro
This patch adds in the needed macro for the ofextensions/circuitsw
project.

Since '/' is a reserved character in Jenkins names (due to directory
structure lay down) we're going to need to create some extra variables
in our higher level scripts to handle META project namespaces.

In particular, we'll need to do appropriate mapping between - and /

This is mostly an issue for the git repo that is pulled.

Change-Id: Ic4502f95a3059fc8557256400e7deceeef2605ef
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>