Add ofextensions/circuitsw project macro
This patch adds in the needed macro for the ofextensions/circuitsw
project. Due to the meta-project name they settings file macro _does_
have a '/' in the name. This is so that our JJB scripts should all work
with no extra configuration as long as the project is identified as
ofextensions/circuitsw. If this is too confusing we may need to consider
adding a METAPROJECT variable to our scripting to deal with such setups
and we can then appropriately concatenate with either a / or - character
where needed.

Change-Id: Ic4502f95a3059fc8557256400e7deceeef2605ef
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>