Update patch set 1
Patch Set 1:

Happy to work with you guys on this.  The last time we talked about this, we were waiting for the transition to red hat/centos 7 for the base VMs.  Have we made that switch yet?  Or should we try to make red hat 6 work?

Patch-set: 1
Reviewer: Gerrit User 46 <46@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=0