Stop triggering on draft contributions to Integration
As sandbox allows to create jobs against changes under review,
it is not essential to run verify jobs on every patch set.

This change makes it so only non-drafts are verified automatically;
so that users interested in Snadbox testing only
have a way to avoid resource consumption on main Releng silo.

Drafts can still be verified using "recheck" and "reverify" comments.

Change-Id: I4b50ad4aae6f05c4f05d4b7141c77ad2b5a281cb
Signed-off-by: Vratko Polak <vrpolak@cisco.com>