Update patch set 1
Patch Set 1: Code-Review-2

Even if I was to consider merging this, I wouldn't want it to be a completely separate docker instance and would want the robot additions added to the current docker definition that we have.

As it stands, I won't merge for a completely new docker definition when there is no good reason for it.

Patch-set: 1
Reviewer: Gerrit User 1 <1@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-2