Update patch set 3
Patch Set 3:

As an aside, and does not need to be done in this patch but I'd like to see us move away from adding these installs here. There is several examples in the jjb folder of how to use virtualenv and pip to install python related dependencies so I want to see integration builders move to using virtualenv and pip in the future so that the vagrant and spin-up scripts become less of a bottleneck for projects to get what they need.

Patch-set: 3