WIP: Integration jobs based on new integration/* repos
Existing jobs work, but currently it is not easy
to develop new tests, as Sandbox would still access
the old read-only repository.

In the long run, this change is needed to be merged,
but for sandbox, this change is useful even before merge.
Feel free to improve.

Few changes at once:
* distribution jobs clone integration/distribution instead integration
* csit jobs clone integration/test instead of integration
* *-bundle-vars.sh used from every csit-like job
* integration/test always use branch master
* various integration macros have project (or even branch) hardcoded
* e-mail publishers with some new tags, needs rework

Change-Id: I9c51b0b702b3c4040c5ed2a3bc1600a6d0113f2c
Signed-off-by: Vratko Polak <vrpolak@cisco.com>