DO NOT MERGE: Increase log size limit for 1node scit jobs
Current limit 20MB is not enough when a scale test hits a bug.
This change increases limit to 10GB.

But, this could consume archive space reqlly quick,
so this change also removes karaf.log archive post-builder.
The intent is not to merge this change,
just use it for Sandbox testing,
where karaf.log is visible in Workspace.

Change-Id: Ib9d544b500542777083eda9e218253a81d1347e3
Signed-off-by: Vratko Polak <vrpolak@cisco.com>