Update patch set 1
Patch Set 1: Code-Review-1

(4 comments)

Ok, I know it seems like I'm basically rewriting your patch, but these are the modifications that I know need to be made for this to be correct. Once this is updated merged there will be another patch I'll make that should finish off the update to get this working correctly.

Patch-set: 1
Label: Code-Review=-1