Update patch set 2
Patch Set 2:

> is it not possible to set log level to ERROR when doing scalability tests?

In principle, it is possible, but it would need a separate change to do enable that.

But it will not help, we could still get tens of thousands of long ERROR messages.

Perhaps if we could disable the rotation behavior, making logging stop when size is reached, we do not need to change the size limit.

Patch-set: 2
Reviewer: Gerrit User 2046 <2046@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0