Update patch set 1
Patch Set 1:

This is one of the consequences of going to single branch (see item 2 at https://lists.opendaylight.org/pipermail/integration-dev/2015-September/004416.html): the single branch stores stable/lithium and master test code, so which distribution do we want to use to verify test code? I think master is better because stable/lithium may not cover new features develop in master. In any case if this is a concern we can bring the discussion to the integration list.

Patch-set: 1