Change MININET to TOOLS_SYSTEM_IP in get-slave-addresses.sh
This change is an attempt to reduce confusion and
increase clarity for testing contributors.
Current MININET variable passed to pybot
(IP address of jenkins slave) refers to any slave node
except CONTROLLER slaves. As testing expands
to include slaves besides mininet, it is prudent
to change VAR MININET to a descriptive name.
This change is backwards compatible with robot scripts
that use global MININET VAR as these variables
are suppled with TOOLS_SYSTEM_IP VAR values on pybot
execution.

Change-Id: I637b877f31f24326aab6ecc2a402d685903ad2c6
Signed-off-by: Marcus G K Williams <Marcus.G.Williams@gmail.com>