Update patch set 2
Patch Set 2:

> This can be done in make-before-break fashion. That means, in the
 > first change the new variable names can be added, but keep the old
 > MININET ones still there. Then merge 26787, then remove the old
 > ones.
 
+1

 > And have we gone through a proper flamewar to establish the winning
 > new name? I still have hopes for TOOLS_VM_IP.

I think it's time for the war :)  Doing it via gerrit is fine with me.  Thanks
for the initiative, Marcus.  To me TEST_PARTNER is not intuitive enough,
and that's the goal here right?  To have more meaningful names.  Also this is
our chance to plan for multiple systems.   So, I'd like to get rid of any 
CONTROLLER variable and have CONTROLLER_00_IP CONTROLLER_01_IP etc.  likewise with
CONTROLLER_00_USER, CONTROLLER_01_PASSWORD.

we can have a DEFAULT_CONTROLLER_USER and DEFAULT_CONTROLLER_PASSWORD that is 
probably always used, but it will give us flexibility if we need it.

As for the PARTNER systems, I don't have a strong opinion on the name(s), but words
like system or vm (vm might be antiquated when our tools are only running in a
container on the same system).  TOOLS_SYSTEM_00_IP, ODL_TOOLS_SYSTEM_00_USER  ???

Patch-set: 2
Label: Verified=0