Update patch set 1
Patch Set 1:

This check is not critical but it is useful to verify all bundles are up after 1 minute the restconf interface becomes avaiable. What is exactly the issue with this and why sshpass is not always available? does the sshpass install fail during VM start up?

Patch-set: 1
Label: Verified=0