Update patch set 1
Patch Set 1:

> Abandoned
 > 
 > The proxy to use is probably written somewhere in settings.xml. My
 > new bash value is completely ignored my Maven.

It's a Jenkins environment variable that we have defined. Those are things that can't be directly overwritten by the jobs IIRC. It's a good thing too, that is a variable we _do_not_ ever want to be overwritten by a job for mechanical reasons. If there are sync issues by the internal proxies we need to be made aware of them.

Patch-set: 1
Reviewer: Gerrit User 1 <1@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0