Add scapy and udpreplay to robot Vagrant image
Scapy is Python library that allows very flexible packet manipulation
and can be used to generate arbitary .pcap files.

The udpreplay utility is a dependency of the lispflowmapping performance
tests, and is used to replay UDP packets from a .pcap file, without
needing root preivileges. In contrast to `tcpreplay`, replaying over the
loopback interface is possible.

Change-Id: Ia55ea21d6a9ce4832c3a14a0454d4341e1f578e2
Signed-off-by: Lorand Jakab <lojakab@cisco.com>