Update patch set 3
Patch Set 3:

> I think this patch will impact the creation of a new generic
 > "mininet" vagrant folder, when do you plan to merge this?

I think the impact is minimal as all I'm doing is refactoring the scripts for code reuse. It doesn't change any actual configuration.

I was talking to Andy today and we are planning on migrating the mininet VMs over to Ubuntu though which I'm looking at now. I think that will be an impact but you will have to switch to it using JJB configuration first so we can test jobs before migration.

Patch-set: 3