Move RobotFramework dependencies to csit job
This work is needed to make maintenance of python modules simplier and
move them out of the vagrant / spinup scripts and directly into the job
itself. This work is necessary to allow us to upgrade to CentOS7 for our
robotframework slave.

Use virtualenv to setup robotframework at the job level. Further
improvement to this could be to use a JJB variable to allow projects to
define which python modules they need in their project file.

Change-Id: Ib167e5ba2989269df177ec070811c83618c25cb6
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>