Update patch set 8
Patch Set 8:

OK, it seems in this OVS version when I remove and add a port, it shows up with the same name which is what I would expect, for some reason it was not working like that in old mininet image. So I can adjust the test (once this is merged) to fix 4 first failures. Now the group and meters is another story, we need CPqD switch for that to work I can try to look at how to build CPqD in Ubuntu 14.04, last time I check they only supported Ubuntu 13 or older version.

Patch-set: 8