Update patch set 5
Patch Set 5:

> Can you reproduce the leak issue outside ODL infra or is it only
 > happening here?

it's only happening here, afaik, and only happening with c7-robot with c7-java.  the
thread leak test is really just checking thread count on java processs (c7-java) then
running the nc tool to open/close a bunch (600) connections on port 6633.  this is 
done local to the controller vm (so, c7-java still).  then it counts the threads
again (c7-java) and if it's too big of a difference, it's a leak  None of that
seems like it should be affected by the c7-robot VM.

but it's consistently happening in the sandbox.

Patch-set: 5