Update patch set 5
Patch Set 5:

> When you say it all happens in the controller, yesterday when I
 > checked the test case I understood the nc tool runs in the robot VM
 > and sends the conn requests to the controller VM. If that is the
 > case I am thinking of some Firewall issue blocking the TCP
 > disconnect sequence. Are all Firewalls off in the 2 VMs?

AH!  you are right.  my bad.  I misread my own suite.  this is where we
need to look then.


so, "nc" is running from the robot vm and pointing at the controller vm.
maybe el6 is blocking those requests and is hiding a real thread leak.
c7 might be letting it through and it exposes the real bug on the controller
side.

I can't do it today, but tomorrow I can take the same distro locally and
see if the thread leak is there.  I know it's something that's shown up
from time to time.

Thanh, do you know if there is anything in el6 that might block SYN packets
egress on TCP port 6633?

thanks Luis!

Patch-set: 5