Update patch set 4
Patch Set 4:

It's my belief now that this failure is valid and the thread leak bug [0] is still there.  Why this is not seen in releng could be the same reason it's
not seen with the el6 robot vm.

I'm hitting this failure locally with both the Lithium release and the latest master distro (from 9/22).

I'll re-open the bug and I would say this failure should not be blocking this 
gerrit any more.

https://bugs.opendaylight.org/show_bug.cgi?id=2429

Patch-set: 4
Label: Verified=0