adding kill to list of commands that can be sudo'd
there is a utility "sshuttle" that is being called with sudo which
opens a tunnel to a remote system.  When the current robot vm is
done with it's job, it will be the same vm to be used in the next
job if there is a job in the queue to use it.

Allowing 'sudo kill' will let jobs using sshuttle to kill off any
old/unwanted ssshuttle processes and restart a new one as needed.

Change-Id: I72f2b1db9a40bb40ee93d8b17a3e9bb77b92f941
Signed-off-by: Jamo Luhrsen <jluhrsen@redhat.com>