Update patch set 1
Patch Set 1:

> > ${WORKSPACE} is already pretty short IMO. Last week we changed
 > > ${WORKSPACE} to be /w/workspace/${JOB_NAME} due to the long
 > script
 > > issue.
 > >
 > > I'm worried about /tmp/${script_name} because of the possibility
 > of
 > > file name collision with other jobs so it should at least be
 > > /tmp/${JOB_NAME}/${script_name} but I think the ${WORKSPACE} is
 > > sufficient already considering we moved it last week.
 > 
 > ok.  then we need another solution for jobs like these:
 > 
 > Luis' patch to change "lithium-redesign" to "li" might be enough,
 > but I'm worried
 > that this problem is going to catch up to us again down the road.
 > 

Isn't this fixed now with your script that moves the pip install into a a separate script? I don't think we need to worry about this problem anymore. The output of the log looks like it was from the pre-script update.

 > As far as conflicts, if we just remove the script after executing
 > it, the next
 > job can do whatever it wants without a chance conflicting.  we
 > don't need this
 > script after it's executed right?

What I'm worried about is if another project just happened to use the same filename as your script in your job but their script is different. yes this chance might be low but it's a possibility.

Patch-set: 1